:art: Rename splitcontent to split-content
