+++ src/components/common/split-content.tsx
