+++ src/pages/design-sprint.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import Layout from '../components/modules/layout'
 import styled from '../util/styled-components'
-import SplitContent from '../components/common/splitcontent'
+import SplitContent from '../components/common/split-content'
 import HeroHeader from '../components/common/hero-header'
 import svgPostIts from '../images/design-sprint-post-its.svg'
 import svgExplanation from '../images/design-sprint-explanation.svg'
